GL-1.1: refactor to adapt to PlatformIO Core v4.1.0 from v3.x. Refactor pin to uint8_t

+++ lib/Bluetooth_Lib/CBluetooth.h
@@ -12,7 +12,7 @@ class CBluetooth
 {
 	public:
 		// Constructors
-		CBluetooth(int iPinRx, int iPinTx, EBluetoothAdapter adapter) : m_Serial(iPinRx, iPinTx)
+		CBluetooth(uint8_t iPinRx, uint8_t iPinTx, EBluetoothAdapter adapter) : m_Serial(iPinRx, iPinTx)
 		{
 			m_adapter = adapter;
 		};
